requirements.txt
# Core dependencies
kubernetes>=28.1.0
prometheus-client>=0.19.0
python-dotenv>=1.0.0
requests>=2.31.0
PyYAML>=6.0.1
langchain>=0.1.0
langgraph>=0.0.15
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0

# Optional: LangGraph for advanced agent coordination
# Uncomment the line below if you want to use LangGraph
# langgraph>=0.0.15

# Monitoring and metrics
prometheus-client>=0.17.1

# Data processing
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0
statsmodels>=0.14.0

# ML and prediction
scikit-learn>=1.3.0
prophet>=1.1.0
tensorflow>=2.12.0  # Optional, for more advanced ML models

# Utilities
pyyaml>=6.0.1
python-dateutil>=2.8.2
tqdm>=4.65.0
joblib>=1.3.0

# LLM
transformers>=4.36.0
torch>=2.1.0
pydantic>=2.5.0
typing-extensions>=4.8.0